pyproject.toml
[project]
name = "ads2u-pipeline"
version = "0.1.0"
description = "Marketing data pipeline for Meta Ads and Google Ads"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.2.0",
    "pandas>=2.0.0",
    "facebook-business>=24.0.0",
    "google-ads>=29.0.0",
    "pyyaml>=6.0",
    "google-auth-oauthlib>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "requests>=2.32.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: unit tests",
    "integration: integration tests",
    "e2e: end-to-end tests requiring Docker running",
]
